before redesign og the addUser
